Corrects indentation for config example (#20)
TL;DR
-----

Fixes the indentation for the KOTS config

Details
-------

Today when Pixee and I tried to use the embedded/external database sample code, realized that we couldn't copy and paste as easily as we wanted. The challenge we faced was that the indentation was off---all of the items had their name entry indented too far. This fix updates the indentation.